Extract auth middleware into a submodule (reclaim guardrail headroom)
Moves the CRITICAL auth path — admin-token auth + RBAC (require_admin_token),
admin-action attribution (record_admin_action_audit + should_record_admin_action),
transport identity (require_client_identity) and transport security
(require_secure_transport), plus the two G7 test modules — out of the 4382-line
bin entry point into a new src/bin/kirra_verifier_service/auth.rs submodule. This
is the #721 startup.rs extraction pattern: `use super::*` inherits the bin's
shared imports (descendant-module visibility, exactly like the route-handler
submodules), and `use auth::*` re-exports the pub(crate) middlewares so the router
assembly references them unqualified.

BEHAVIOUR-PRESERVING: the functions are moved byte-identically; the same 82 binary
tests pass (the moved G7 tests now namespaced auth::g7_*). The only reference
change is the router test's `use super::build_app` -> `use crate::build_app`
(build_app stays at the crate root). CRITICAL INVARIANTS #1/#2/#6/#13 are
unchanged — this is a pure relocation.

Guardrail: the bin TIGHTENS 4382->4079 lines and 126->121 panics (the five test
expect/unwraps moved with the router tests) to lock the gain; auth.rs is newly
tracked at 328 lines / panic_budget 5 (all five in the moved router tests — the
production auth code is panic-free) and added to ownership_scope.

cargo test --bin: 82 pass; clippy clean; ci/check_quality_guardrails.py passes.
