+++ ci/quality_guardrails_baseline.json
@@ -1,14 +1,16 @@
 {
-  "_comment": "Quality-guardrail ratchet baseline. Metrics may only DECREASE without review; an INCREASE fails CI (see ci/check_quality_guardrails.py). #721 (startup.rs extraction) then merged, removing ~227 lines from the bin \u2014 bin line ceiling tightened 4186\u21923959 to lock that gain, and the new SG-008 startup.rs is now tracked (line ceiling + panic_budget 0, keeping the fail-closed predicate panic-free) and added to ownership_scope. WS-0.5 (/metrics, PR #774, reviewed+merged) grew three tracked files with observability code and its tests: the bin gained the LockedOut scrape DoD test (3997\u21924087 lines; panic budget 117\u2192121 \u2014 the four new expect/unwraps are all in that test), verifier.rs gained the AppState fleet_metrics field (1092\u21921099), policy_layer.rs gained the denial classifier + counting + two test modules (1036\u21921154). Fix-PR D (#772 F3, audit-durability observability) added the `incident_durability_failures` counter: verifier.rs gained the AppState field + doc (1099\u21921108), and the bin's LockedOut-scrape DoD test gained one expected-series line (4087\u21924088). No new panics in either file (the counter field is not a panic site). Fix-PR E (#774 F1, silent scrape resolver) added `resolve_posture_snapshot_silent` to the bin's posture_engine_v2 import, expanding the single-line `use` to three lines (4088\u21924090); no new panics. (policy_layer.rs shrank \u2014 `classify_gate_denial` was deleted in favour of the authoritative `route_command_verdict` \u2014 so its ceiling stays.) Ceilings re-pinned at exactly those values \u2014 the next growth still fails. When a refactor legitimately raises a ceiling, bump it with a justification in the PR; when it lowers one, tighten it here. G7 key/identity lifecycle (per-principal tokens #802, RBAC #803, attribution #804) grew the bin with the auth + record-admin-action attribution middleware and their tests (4090\u21924241, no new panics). G7 transport-security enforcement (#805, the mesh-mTLS option) grew the bin with the `require_secure_transport` middleware + the router-level ordering tests (4241\u21924382; panic 121\u2192126 \u2014 the five new expect/unwraps are all in the router test module) and grew verifier.rs with `TransportSecurityConfig` + the pure `request_transport_is_secure` predicate + its unit tests (1108\u21921233; panic 27\u219228 \u2014 the one new unwrap is a `HeaderValue::from_str` in a test).",
+  "_comment": "Quality-guardrail ratchet baseline. Metrics may only DECREASE without review; an INCREASE fails CI (see ci/check_quality_guardrails.py). #721 (startup.rs extraction) then merged, removing ~227 lines from the bin \u2014 bin line ceiling tightened 4186\u21923959 to lock that gain, and the new SG-008 startup.rs is now tracked (line ceiling + panic_budget 0, keeping the fail-closed predicate panic-free) and added to ownership_scope. WS-0.5 (/metrics, PR #774, reviewed+merged) grew three tracked files with observability code and its tests: the bin gained the LockedOut scrape DoD test (3997\u21924087 lines; panic budget 117\u2192121 \u2014 the four new expect/unwraps are all in that test), verifier.rs gained the AppState fleet_metrics field (1092\u21921099), policy_layer.rs gained the denial classifier + counting + two test modules (1036\u21921154). Fix-PR D (#772 F3, audit-durability observability) added the `incident_durability_failures` counter: verifier.rs gained the AppState field + doc (1099\u21921108), and the bin's LockedOut-scrape DoD test gained one expected-series line (4087\u21924088). No new panics in either file (the counter field is not a panic site). Fix-PR E (#774 F1, silent scrape resolver) added `resolve_posture_snapshot_silent` to the bin's posture_engine_v2 import, expanding the single-line `use` to three lines (4088\u21924090); no new panics. (policy_layer.rs shrank \u2014 `classify_gate_denial` was deleted in favour of the authoritative `route_command_verdict` \u2014 so its ceiling stays.) Ceilings re-pinned at exactly those values \u2014 the next growth still fails. When a refactor legitimately raises a ceiling, bump it with a justification in the PR; when it lowers one, tighten it here. G7 key/identity lifecycle (per-principal tokens #802, RBAC #803, attribution #804) grew the bin with the auth + record-admin-action attribution middleware and their tests (4090\u21924241, no new panics). G7 transport-security enforcement (#805, the mesh-mTLS option) grew the bin with the `require_secure_transport` middleware + the router-level ordering tests (4241\u21924382; panic 121\u2192126 \u2014 the five new expect/unwraps are all in the router test module) and grew verifier.rs with `TransportSecurityConfig` + the pure `request_transport_is_secure` predicate + its unit tests (1108\u21921233; panic 27\u219228 \u2014 the one new unwrap is a `HeaderValue::from_str` in a test). Auth-middleware extraction (#721 pattern) then moved the auth path — admin-token auth + RBAC, transport identity/security, admin-action attribution, and their G7 tests — out of the bin into the new tracked `auth.rs` submodule: the bin TIGHTENS 4382→4079 lines and 126→121 panics (the five test expect/unwraps moved with the router tests) to lock the gain; `auth.rs` is newly tracked at 328 lines / panic_budget 5 (all five are in the moved router tests — the production auth code is panic-free) and added to ownership_scope.",
   "max_lines": {
-    "src/bin/kirra_verifier_service.rs": 4382,
+    "src/bin/kirra_verifier_service.rs": 4079,
+    "src/bin/kirra_verifier_service/auth.rs": 328,
     "src/bin/kirra_verifier_service/startup.rs": 248,
     "src/gateway/mod.rs": 538,
     "src/verifier.rs": 1233,
     "src/gateway/policy_layer.rs": 1154
   },
   "panic_budget": {
-    "src/bin/kirra_verifier_service.rs": 126,
+    "src/bin/kirra_verifier_service.rs": 121,
+    "src/bin/kirra_verifier_service/auth.rs": 5,
     "src/bin/kirra_verifier_service/startup.rs": 0,
     "src/gateway/mod.rs": 6,
     "src/verifier.rs": 28,
@@ -16,6 +18,7 @@
   },
   "ownership_scope": [
     "src/bin/kirra_verifier_service.rs",
+    "src/bin/kirra_verifier_service/auth.rs",
     "src/bin/kirra_verifier_service/startup.rs",
     "src/bin/kirra_verifier_service/attestation.rs",
     "src/bin/kirra_verifier_service/fleet.rs",
